+++ packages/fighting-design/expand-card/src/interface.d.ts
@@ -11,4 +11,7 @@ export interface ExpandCardImageListItem {
   text?: string
 }
 
+/**
+ * 掺入的图片列表类型接口
+ */
 export type ExpandCardImageList = | ExpandCardImageListItem[] | string[]