+++ CHANGELOG.md
@@ -19,6 +19,7 @@
 - `f-aside` 组件 `width` 默认值改为 null
 - `f-image` 组件 `round` 默认值改为 null
 - `f-image-preview` 组件 `z-index` 默认值改为 null
+- `f-expand-card` 组件 `color` 默认值改为 null
 
 ## 0.9.0 (2022-12-07)